cpp: fixes for unique_ptr and pclose interaction
Summary:
The glibc and/or libstdcpp found in the latest upstream
lf-master refresh does not like using `pclose` to a `unique_ptr`
directly because the return value of the `pclose` is ignored.  Instead,
we can wrap it in a simple lambda so that the return is explicitly
ignored (by casting to `void`).

This is the same changes made upstream for similar code such as:
    https://github.com/openbmc/phosphor-fan-presence/commit/fc2b078ac2969b24c12b22a9d522e5bc73760c8a

Test Plan: Expecting CIT compile to pass; code is roughly equivalent.

Reviewed By: wangx6f

Differential Revision: D54598090

fbshipit-source-id: 8ff53c9fbe23c23ce2a2985a8c0d2cabe2181e32