yosemite4: Adjust ina233 config for matching hardware design (#2997)
Summary:
# Description:
Adjust ina233 config for matching hardware design. The external shunt is 5 mohm, so the resistor-calibration should be 0x0400.

This patch is for hardware testing. According to the datasheet and upstream review comment, resistor-calibration should be calculated, and we should set external shunt on DTS instead. We will revise the ina233 driver by the upstream review comment and contribute the code once we finish.

# Motivation:
Hardware team reported that the current reading wasn't correct.

Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/2997

Test Plan: Measure the current with current load and compare with the BMC's reading - pass

Reviewed By: wangx6f

Differential Revision: D54569359

fbshipit-source-id: 134dfeb4881c93013e05469631c47ded1de0ad70