Meru: Upstreaming support for P4 Fairywren CPLD upgrades
Summary:
# Description
This PR enables support for P4 Fairywren CPLD upgrades.

In addition, the latest P1 and P4 CPLD images have changed the upgrade procedure somewhat
to not turn off the CPU prior to upgrades since there is no good reason for turning it off.

# Motivation

This change is needed in order to support P4 Fairywren CPLD upgrades.

X-link: https://github.com/facebookexternal/openbmc.arista/pull/350

Test Plan:
On a P4 Fairywren, used the fpga_util.sh utility to verify that the CPLD could
be upgraded. Verified that wedge_util.sh was able to power on and off the CPU
and verified the CPU booted.

Repeated this test on a P1 Fairywren.

Reviewed By: somasun

Differential Revision: D54736190

Pulled By: joancaneus

fbshipit-source-id: 604b7b102262c61213be2ff94740a135093ea0c9