fix REST API on lf-master
Summary:
Rest API doesn't start on the latest meru image, which blocks BMC
configuration and thereby provisioning.

```
2024 Mar 19 06:57:51 fboss323770036-oob.snc1.facebook.com daemon.info meru-2024.11.1: rest.py[8531]:     from .client import (
2024 Mar 19 06:57:51 fboss323770036-oob.snc1.facebook.com daemon.info meru-2024.11.1: rest.py[8531]:   File "/usr/lib/python3.12/site-packages/aiohttp/client.py", line 34, in <module>
2024 Mar 19 06:57:51 fboss323770036-oob.snc1.facebook.com daemon.info meru-2024.11.1: rest.py[8531]:     import attr
2024 Mar 19 06:57:51 fboss323770036-oob.snc1.facebook.com daemon.info meru-2024.11.1: rest.py[8531]: ModuleNotFoundError: No module named 'attr'
```

Add a dependency on the python attr module to get rest.py to start.
This should be part of the aiohttp module but I think we'd need to contribute
that upstream to Yocto which will take a while to filter through.

Test Plan:
Build meru image on devserver and boot in qemu (as elbert machine).  Observe
that Python attr error is no longer thrown.

Reviewed By: kawmarco

Differential Revision: D55132668

fbshipit-source-id: 9a73a67d40f4a4f3b2b0d4eb10defa2938a80e12