don't decode VLAN tagged packets
Summary:
This is blocking an NPI because it's unexpectedly picking up trunked
LLDP packets (diagnosis from hrvoje).

Handling VLANs properly is tricky (e.g. recursion) and we don't have a use
case internally for it at Meta.  When we do, it might be more useful to move
to a 3rd party tool that does this better.  In the meantime, you can run
lldp-util against a VLAN interface if you need to see what's going on
(e.g. lldp-util -i eth0.4088).  #movefast #efficiency

Test Plan: Green signals in Phabricator before landing.

Reviewed By: pallotron

Differential Revision: D55151985

fbshipit-source-id: 79395084a11ec6cace8df2683a3bf36a586d2fb6