fruid-util: switch prints to stderr
Summary:
The `fruid-util` is often used with a JSON output for tooling
purposes.  If the tool encounters an error and uses `printf` this
results in invalid JSON on the output, which has to be post-processed to
remove.  Fix all the error prints to use `stderr` instead of `stdout`.

Test Plan: Compile passes.  Trivial changes.

Reviewed By: doranand

Differential Revision: D55604889

fbshipit-source-id: 9d230454fdf267215cc2215579f72a62a42c7bb5