GTI: UBB: fix GPU FW empty after GPU FW update
Summary:
# Description
When we perform an AC cycle after update GPU FW, UBB will take some time to activate GPU components. It results in BMC reads an empty string and stores it in the cache. In this PR, We ignore the empty string to avoid this issue.

X-link: https://github.com/facebookexternal/openbmc.quanta/pull/4274

Test Plan: Update GPU FW and check if FW version is correct

Reviewed By: williamspatrick

Differential Revision: D57327444

fbshipit-source-id: eed630725784aeee1355424f5da6fa3fd4c49f46