restapi: Fix missing file in .bb
Summary:
D57500421 added a new file but did not add it to the
corresponding .bb file. As a result, the file was not present when
flashing a new image and the REST service fails to start.

Test Plan: flash onto a box from basset and see that REST service is up

Reviewed By: srikrishnagopu

Differential Revision: D57654158

fbshipit-source-id: 45f6848b500f6597f2edfa5a593dbb3df33a8a32