fuji: Increase dual boot watchdog timeout 5mins -> 10mins
Summary:
This is a prerequisite to enlarging the data0 partition size,
which causes a format during upgrade, which takes > 5mins.

Test Plan:
```
/data/users/alandau/openbmc/build-fuji$ objdump -d tmp/work/fuji-fb-linux-gnueabi/u-boot/v2019.04/build/u-boot

0000327c <fmc_enable_dual_boot>:
...
    32c4:       e3012770        movw    r2, #6000       ; 0x1770
    32c8:       e5832068        str     r2, [r3, #104]  ; 0x68
```

On fuji, the dual boot watchdog is enabled through FMC. The code above
shows that we're writing 6000 to the timeout register, which is 10 mins
in units of 0.1 sec.

Reviewed By: Scott8440

Differential Revision: D57647867

fbshipit-source-id: 4ad5690a090f05e2b2daba6b7675cbc08dc4bc30