minipack libpal: Fix zombie processes and LEDs not blinking
Summary:
The old code was running out of memory sue to a bug in memory
allocation, which leads to not closing the `popen`-ed `ps w` process,
which leaks file descriptors and prevents zombie `ps` processes from
being reaped. In addition, LED statuses are not being updated correctly
in this situation.

This diff changes the logic to avoid allocating memory for the entire
`ps w` output. Rather it reads the output line by line, searches for the
correct processes, and updates the correcponding LED flags.

Test Plan:
Count the number of zombie `ps` processes with
`ps w | grep '\[ps\]' | wc -l`.

Before this diff, the number increased by 1 every ~20s up to 1018, which
is close enough to the default limit on file descriptors (1024) to be
the likely culprit.

After, always 0.

Reviewed By: doranand

Differential Revision: D58115211

fbshipit-source-id: fcc0eccd2af26b0c0a238ce8345a36b91a5094ad