mfgtool: inventory: fix Slot display
Summary:
The `Decorator.Slot::SlotNumber` value was being
displayed as a boolean rather than a number.  Digging into
this it is because the entity-manager representation for
numbers does not match the dbus interface specification for
SlotNumber (suppose to be a `size` but EM is using `uint64`).

Work around this by adding `uint64_t` to the possible types
in the variant read from dbus.  Since `size_t` is already a
type, we cannot use the normal `std::variant` anymore because
on 64-bit architectures there would be a duplicate type in
the variant definition.  sdbusplus has a type called
`dedup_variant` that allows duplicate types, so switch to
that.

Test Plan:
```
    "system/board/Yosemite_4_Wailua_Falls_Slot_8": {
        "Decorator.Asset": {
            "Manufacturer": "Wiwynn",
            "Model": "Yosemite V4 EVT",
            "PartNumber": "N/A",
            "SerialNumber": "N/A",
            "SparePartNumber": "19-100101"
        },
        "Decorator.AssetTag": {
            "AssetTag": "N/A"
        },
        "Decorator.Revision": {
            "Version": "EVT"
        },
        "Decorator.Slot": {
            "SlotNumber": 8
        },
        "Item.Board": {
            "Name": "Yosemite 4 Wailua Falls Slot 8",
            "Type": "Board"
        }
    }
```

Reviewed By: wangx6f

Differential Revision: D58744309

fbshipit-source-id: 62a65a9a432e85817a4b2454ca03899184f9941c