rackmon2: Add helper scripts to aid in register map creation
Summary:
Adds scripts to convert register map from/to CSV.

The primary reason is to make writing the exceptionally large
register maps easy. The user could do all the writing in
a spreadsheet and then have it converted into a JSON in
a couple easy steps with a lot less errors.

Test Plan:
Convert a known map to CSV
```
$ ./regmap_format.py ../configs/register_map/orv3_psu.json
$ cat a.csv | pastry
```
P1425652053

Convert it back to JSON
```
$ ./regmap_format.py ./a.csv
$ cat a.json | pastry
```
P1425652871

Diff them and make sure there are no differences in the register
map ("registers" key).
```
$ diff <(jq < ../configs/register_map/orv2_psu.json) <(jq < ./a.json) | pastry
```
P1425653956

Reviewed By: jagpalgill

Differential Revision: D58495256

fbshipit-source-id: 5c2d943a7fba0e2d2a815cb464ecdd4d366d2491