rackmon2: Allow register map to specify if value has sign
Summary:
This allows us to explicitly mark certain registers as
signed or unsigned (default) thus allowing the core to
do any appropriate sign extension on registers and ensure
that they will be decoded as negative instead of large
positive numbers.

A default of unsigned (sign = false) was chosen after
consulting the register maps of the various devices and
confirming that a majority are recommended to be unsigned.

Test Plan:
Add unit tests to make sure we treat numbers appropriately.
(0xfffe is treated as unsigned if register descripter has
the sign flag set to false). Ensure we are decoding as unsigned
by default when not set explicitly.

Reviewed By: jagpalgill

Differential Revision: D58690175

fbshipit-source-id: def4b6b23f21de6c01f79768d9e932168b2e321b