common: ssd-mond: Monitor hot-swappable SSD presence
Summary:
# Description

This pull request introduces a new feature to the common layer to monitor the presence of a hot-swappable SSD.

# Motivation

As the SSD in newer platforms like janga and tahan can be easily removed and inserted, this feature generates BMC logs and ipmi event logs on COMe side when the SSD presence changes. As the presence of SSD cannot be monitored from the COMe this calls for monitoring a exported CPLD register bit(indicating the SSD presence) as a sysfs variable from the BMC.
The sysfs variable shoudl be of the format:
```
root@bmc-oob:~# head /sys/bus/i2c/devices/1-0035/e1_ssd_present
0x1

Note:
 0x00: absent
 0x01: present

Bit[0:0] @ register 0x27, register value 0xff
```
This feature adds log entries to the syslog in case of presence change, which can be checked from the BMC side, and once the COMe is reset, we can print the ipmi event list to check the device presence event logs.

X-link: https://github.com/facebookexternal/openbmc.celestica/pull/1774

Test Plan:
Tested ssd-mond's functionality on systemd platforms:

**SystemD platform:**

1. Built the BMC image for Tahan platform after adding changes from the PR:
2. The functionality is validated by physically removing and inserting an SSD, worked as expected.

**Note:** This feature is not supported on sysvinit platform.

Reviewed By: mikechoifb

Differential Revision: D58936778

Pulled By: alandau

fbshipit-source-id: 1625aacaf54a1d5faa9dd3f33895c1f40a4dd0bb