rackmon2: Add update support for regmap format scripts
Summary:
Add functionality for the script to update the output
JSON instead of overwriting it with stock values for name/etc.

Considering how easy it is to edit the CSV files as opposed
to JSON, it would be nice if we could have the script
update an existing JSON instead of overwriting it.

Test Plan:
regmap_format.py csv/orv3_psu.csv -o orv4_psu.json --update
and ensure none of the other fields not covered by the CSV
are affected.

Reviewed By: jagpalgill

Differential Revision: D59530892

fbshipit-source-id: 4a4c0db81035015e386c34c9b10b91c4040dd12d